requirements.txt
Flask==2.3.3
paramiko==3.4.0
mysql-connector-python==8.2.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
cryptography==41.0.7